+++ pipeline/builder/builder_test.go
@@ -81,7 +81,7 @@ func TestBuild(t *testing.T) {
 					Metric:          "int-metric",
 					IntervalSeconds: 10,
 					Value: metrics.MetricValue{
-						IntValue: 10,
+						Int64Value: 10,
 					},
 					Labels: map[string]string{"foo": "bar"},
 				},
